pyproject.toml
[tool.poetry]
name = "duffbot"
version = "0.1.0"
description = "AI Tool Preference Collector"
authors = ["chema <chema@montevive.ai>"]

[tool.poetry.dependencies]
python = "^3.12,<3.13"
fastapi = "^0.115.0"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
gradio = "^4.44.0"
requests = "^2.26.0"
llama-index = "^0.11.11"
python-dotenv = "^0.19.0"
setuptools = "^75.1.0"
jinja2 = "^3.1.4"
llama-index-callbacks-arize-phoenix = "^0.2.1"
redis = "^5.0.8"
pytest-asyncio = "^0.24.0"
[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
black = "^21.9b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
